requirements.txt
comtypes==1.1.11
MouseInfo==0.1.3
mss==6.1.0
numpy==1.22.4
opencv-python==4.5.5.64
Pillow==9.1.0
pip==21.3.1
PyAutoGUI==0.9.53
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.28
pytweening==1.0.4
setuptools==60.2.0
wheel==0.37.1
